User powershell to do more and better setup in windows.

+++ setup.sh
@@ -1,11 +1,11 @@
 #!/bin/sh
 
-echo "Downloading git-completion & git-prompt..."
+echo "Checking for or downloading git-completion & git-prompt..."
 
-curl -o git/git-completion.bash --progress-bar \
-    https://raw.githubusercontent.com/git/git/master/contrib/completion/git-completion.bash
+[ -e "git/git-completion.bash" ] ||  curl -o git/git-completion.bash --progress-bar \
+	https://raw.githubusercontent.com/git/git/master/contrib/completion/git-completion.bash
 
-curl -o git/git-prompt.sh --progress-bar \
-    https://raw.githubusercontent.com/git/git/master/contrib/completion/git-prompt.sh
+[ -e "git/git-prompt.sh" ] || curl -o git/git-prompt.sh --progress-bar \
+	https://raw.githubusercontent.com/git/git/master/contrib/completion/git-prompt.sh
 
 echo "Done."
